@@ -2,53 +2,77 @@
   "name": "rcongui",
   "version": "0.1.0",
   "private": true,
+  "scripts": {
+    "start": "vite",
+    "dev": "vite dev",
+    "dev:public": "REACT_APP_PUBLIC_BUILD=true vite dev",
+    "build": "vite build",
+    "build:public": "REACT_APP_PUBLIC_BUILD=true vite build",
+    "demo": "NODE_ENV=development vite build && vite preview --host",
+    "serve": "vite preview",
+    "preview": "vite build && vite preview",
+    "test": "jest",
+    "format:write": "prettier src/ --write",
+    "lint": "eslint src/"
+  },
   "dependencies": {
-    "@date-io/moment": "^1.3.13",
-    "@fortawesome/fontawesome-svg-core": "^1.2.30",
-    "@fortawesome/free-brands-svg-icons": "^5.12.1",
-    "@fortawesome/free-solid-svg-icons": "^5.14.0",
-    "@fortawesome/react-fontawesome": "^0.1.8",
-    "@material-ui/core": "4.11.2",
-    "@material-ui/icons": "^4.5.1",
-    "@material-ui/lab": "^4.0.0-alpha.45",
-    "@material-ui/pickers": "^3.2.10",
+    "@dnd-kit/core": "^6.3.1",
+    "@dnd-kit/modifiers": "^9.0.0",
+    "@dnd-kit/sortable": "^10.0.0",
+    "@dnd-kit/utilities": "^3.2.2",
+    "@emoji-mart/data": "^1.2.1",
+    "@emoji-mart/react": "^1.1.1",
+    "@fontsource/montserrat": "^5.1.0",
+    "@fontsource/roboto": "^5.0.14",
+    "@fortawesome/free-brands-svg-icons": "^6.6.0",
+    "@fortawesome/free-solid-svg-icons": "^6.6.0",
+    "@fortawesome/react-fontawesome": "^0.2.2",
+    "@jsonforms/core": "^3.4.1",
+    "@jsonforms/material-renderers": "^3.4.1",
+    "@jsonforms/react": "^3.4.1",
     "@monaco-editor/react": "^4.2.2",
+    "@mui/icons-material": "^6.0.2",
+    "@mui/lab": "^6.0.0-beta.9",
+    "@mui/material": "^6.0.2",
+    "@mui/styles": "^6.0.2",
+    "@mui/system": "^6.0.2",
+    "@mui/x-charts": "^7.16.0",
+    "@mui/x-data-grid": "^7.16.0",
+    "@mui/x-date-pickers": "^7.16.0",
+    "@tanstack/react-query": "^5.55.0",
+    "@tanstack/react-query-devtools": "^5.59.15",
+    "@tanstack/react-table": "^8.20.5",
+    "chart.js": "^3.9.1",
+    "country-list": "^2.2.0",
+    "dayjs": "^1.11.10",
+    "emoji-mart": "^5.6.0",
+    "emoji-picker-react": "^4.12.0",
+    "google-palette": "^1.1.0",
+    "immutable": "^5.0.3",
+    "localforage": "^1.10.0",
+    "lodash": "^4.17.21",
+    "moment": "^2.29.4",
+    "react": "^18.3.1",
+    "react-chartjs-2": "^4.0.0",
+    "react-code-blocks": "^0.1.6",
+    "react-copy-to-clipboard": "^5.1.0",
+    "react-dom": "^18.3.1",
+    "react-hook-form": "^7.53.0",
+    "react-router-dom": "^6.26.1",
+    "react-toastify": "^10.0.5",
+    "react-window": "^1.8.10",
+    "tiny-invariant": "^1.2.0",
+    "zustand": "^5.0.0"
+  },
+  "devDependencies": {
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
-    "chart.js": "^3.9.1",
-    "chartjs-adapter-moment": "^1.0.0",
-    "clsx": "^2.1.1",
-    "country-list": "^2.2.0",
-    "emoji-mart": "^3.0.0",
-    "faker": "^4.1.0",
-    "google-palette": "^1.1.0",
-    "immutable": "^4.0.0-rc.12",
-    "lodash": "^4.17.19",
-    "material-ui": "^0.20.2",
-    "moment": "^2.29.4",
-    "monaco-editor": "^0.28.1",
-    "mui-datatables": "^3.7.4",
-    "react": "^16.12.0",
-    "react-beautiful-dnd": "^13.1.1",
-    "react-chartjs-2": "^4.0.0",
-    "react-code-blocks": "^0.1.4",
-    "react-copy-to-clipboard": "^5.0.3",
-    "react-dom": "^16.12.0",
-    "react-router-dom": "^5.1.2",
-    "react-toastify": "^5.5.0",
-    "recompose": "^0.30.0"
-  },
-  "devDependencies": {
-    "@vitejs/plugin-react": "^4.0.0",
+    "@types/lodash": "^4.17.13",
+    "@types/react-window": "^1.8.8",
+    "@vitejs/plugin-react": "^4.2.1",
     "prettier": "^2.8.8",
-    "vite": "^4.5.5"
-  },
-  "scripts": {
-    "start": "vite",
-    "build": "vite build",
-    "serve": "vite preview",
-    "test": "jest"
+    "vite": "^5.4.3"
   },
   "browserslist": {
     "production": [
