@@ -1,18 +1,18 @@
 import {
-  Grid,
   IconButton,
   InputLabel,
   Paper,
   Tooltip,
   Typography,
-} from "@material-ui/core";
-import React from "react";
-import EditIcon from "@material-ui/icons/Edit";
-import DeleteIcon from "@material-ui/icons/Delete";
+} from "@mui/material";
+import EditIcon from "@mui/icons-material/Edit";
+import DeleteIcon from "@mui/icons-material/Delete";
 import { SYNC_METHODS } from "./BlacklistListCreateDialog";
+import Grid from "@mui/material/Grid2";
+import {Fragment} from "react";
+
 
 const BlacklistListTile = ({
-  classes,
   blacklist,
   servers,
   onEdit,
@@ -24,63 +24,65 @@ const BlacklistListTile = ({
   }
 
   return (
-    <Paper>
+    (<Paper>
       <Grid container
-        justify="space-between"
+        justifyContent="space-between"
         alignItems="center"
         spacing={2}
-        className={classes.padding}
         style={{paddingLeft: 18}}
       >
-        <Grid item>
+        <Grid>
           <Typography variant="h6" color="textSecondary">#{blacklist.id}</Typography>
         </Grid>
-        <Grid item xs={2}>
+        <Grid size={2}>
           <InputLabel align="left">Blacklist Name</InputLabel>
           <Typography align="left">
             {blacklist.name}
           </Typography>
         </Grid>
 
-        <Grid item xs={6}>
+        <Grid size={6}>
           <InputLabel align="left">Servers</InputLabel>
           {
             blacklist.servers === null
             ? <Typography align="left">All</Typography>
             : blacklist.servers.length === 0
             ? <Typography align="left">None</Typography>
-            : <React.Fragment>
+            : <Fragment>
               {blacklist.servers.map((num) => (
                 <Typography key={num} align="left" noWrap>
                   {serverNumberToName(num)}
                 </Typography>
               ))}
-            </React.Fragment>
+            </Fragment>
           }
         </Grid>
         
-        <Grid item xs={2}>
+        <Grid size={2}>
           <InputLabel align="left">Sync Method</InputLabel>
           <Typography align="left">{SYNC_METHODS[blacklist.sync]}</Typography>
         </Grid>
         
-        <Grid item>
+        <Grid>
           <Tooltip title="Edit">
-            <IconButton onClick={() => onEdit(blacklist)}>
+            <IconButton onClick={() => onEdit(blacklist)} size="large">
               <EditIcon/>
             </IconButton>
           </Tooltip>
           <Tooltip title="Delete">
             <span>
-              <IconButton onClick={() => onDelete(blacklist)} disabled={blacklist.id === 0}>
+              <IconButton
+                onClick={() => onDelete(blacklist)}
+                disabled={blacklist.id === 0}
+                size="large">
                 <DeleteIcon/>
               </IconButton>
             </span>
           </Tooltip>
         </Grid>
       </Grid>
-    </Paper>
-  )
+    </Paper>)
+  );
 }
 
 export default BlacklistListTile;
