@@ -1,9 +1,10 @@
-import { ButtonGroup, Button, Grid } from "@material-ui/core";
-import React from "react";
-import Tooltip from "@material-ui/core/Tooltip";
-import EditIcon from "@material-ui/icons/Edit";
-import TimerOffIcon from "@material-ui/icons/TimerOff";
-import DeleteIcon from "@material-ui/icons/Delete";
+import { ButtonGroup, Button } from "@mui/material";
+import Tooltip from "@mui/material/Tooltip";
+import EditIcon from "@mui/icons-material/Edit";
+import TimerOffIcon from "@mui/icons-material/TimerOff";
+import DeleteIcon from "@mui/icons-material/Delete";
+import Grid from "@mui/material/Grid2";
+
 
 const BlacklistRecordActionRow = ({
   isExpired,
@@ -12,8 +13,8 @@ const BlacklistRecordActionRow = ({
   onDelete,
 }) => {
   return (
-    <Grid container justify="center">
-      <Grid item>
+    (<Grid container justifyContent="center">
+      <Grid>
         <ButtonGroup size="small" variant="text">
           <Button>
             <Tooltip
@@ -37,7 +38,7 @@ const BlacklistRecordActionRow = ({
           </Button>
         </ButtonGroup>
       </Grid>
-    </Grid>
+    </Grid>)
   );
 };
 
