@@ -1,29 +1,28 @@
-import { Button, FormControl, FormControlLabel, Grid, InputLabel, MenuItem, Select, Switch, TextField } from "@material-ui/core";
-import React from "react";
-import { get, handle_http_errors, showResponse } from "../../utils/fetchUtils";
+import { Button, FormControl, FormControlLabel, InputLabel, MenuItem, Select, Switch, TextField } from "@mui/material";
+import Grid from "@mui/material/Grid2";
+import {useState} from "react";
 
 const BlacklistRecordsSearch = ({
-  classes,
   blacklists,
   onSearch,
   disabled,
 }) => {
-  const [playerIdQuery, setPlayerIdQuery] = React.useState("");
-  const [reasonQuery, setReasonQuery] = React.useState("");
-  const [blacklistQuery, setBlacklistQuery] = React.useState("");
-  const [excludeExpired, setExcludeExpired] = React.useState(false);
-  const [pageSize, setPageSize] = React.useState(50);
+  const [playerIdQuery, setPlayerIdQuery] = useState("");
+  const [reasonQuery, setReasonQuery] = useState("");
+  const [blacklistQuery, setBlacklistQuery] = useState("");
+  const [excludeExpired, setExcludeExpired] = useState(false);
+  const [pageSize, setPageSize] = useState(50);
 
   return (
-    <form className={classes.flexContainer}>
+    (<form>
       <Grid
         container
         spacing={2}
         alignContent="center"
         alignItems="center"
-        justify="space-evenly"
+        justifyContent="space-evenly"
       >
-        <Grid item xs={4}>
+        <Grid size={4}>
           <TextField
             fullWidth
             label="Search by player ID"
@@ -31,7 +30,7 @@ const BlacklistRecordsSearch = ({
             onChange={(e) => setPlayerIdQuery(e.target.value)}
           />
         </Grid>
-        <Grid item xs={8}>
+        <Grid size={8}>
           <TextField
             fullWidth
             label="Search by name or reason"
@@ -39,7 +38,7 @@ const BlacklistRecordsSearch = ({
             onChange={(e) => setReasonQuery(e.target.value)}
           />
         </Grid>
-        <Grid item xs={4}>
+        <Grid size={4}>
           <FormControl fullWidth>
             <InputLabel>Blacklist</InputLabel>
             <Select
@@ -53,7 +52,7 @@ const BlacklistRecordsSearch = ({
             </Select>
           </FormControl>
         </Grid>
-        <Grid item xs={3}>
+        <Grid size={3}>
           <FormControlLabel
             control={
               <Switch
@@ -66,7 +65,7 @@ const BlacklistRecordsSearch = ({
             labelPlacement="top"
           />
         </Grid>
-        <Grid item xs={2}>
+        <Grid size={2}>
           <FormControl fullWidth>
             <InputLabel>Page size</InputLabel>
             <Select
@@ -85,7 +84,7 @@ const BlacklistRecordsSearch = ({
             </Select>
           </FormControl>
         </Grid>
-        <Grid item xs={3}>
+        <Grid size={3}>
           <Button
             type="submit"
             disabled={disabled}
@@ -107,7 +106,7 @@ const BlacklistRecordsSearch = ({
           </Button>
         </Grid>
       </Grid>
-    </form>
+    </form>)
   );
 }
 
