@@ -1,139 +1,101 @@
-import React, { useEffect, useRef, useState } from "react";
-import Badge from "@material-ui/core/Badge";
-import { Comment, Send } from "@material-ui/icons";
-import {
-  Box,
-  Button,
-  Chip,
-  Drawer,
-  Grid,
-  makeStyles,
-  TextField,
-} from "@material-ui/core";
-import IconButton from "@material-ui/core/IconButton";
-import Typography from "@material-ui/core/Typography";
+import {useState} from "react";
+import Badge from "@mui/material/Badge";
+import {Comment, Send} from "@mui/icons-material";
+import {Box, Button, Chip, Drawer, TextField,} from "@mui/material";
+import IconButton from "@mui/material/IconButton";
+import Typography from "@mui/material/Typography";
 import moment from "moment";
+import Grid from "@mui/material/Grid2";
 
-const useStyles = makeStyles((theme) => ({
-  message: {
-    whiteSpace: "pre-wrap",
-    marginTop: "5px",
-    marginBottom: "5px",
-  },
-  date: {
-    color: theme.palette.disabledColor,
-  },
-  padding: {
-    padding: theme.spacing(1),
-    overflow: "auto",
-    maxHeight: "68vh",
-  },
-}));
+const ChatContent = ({data, handleMessageSend}) => {
+  const [comment, setComment] = useState("");
 
-const AlwaysScrollToBottom = () => {
-  const elementRef = useRef();
-  useEffect(() => elementRef.current.scrollIntoView());
-  return <div ref={elementRef} />;
-};
+  return (<Box paddingLeft={2}>
+      <Grid
+        container
+        justifyContent="flex-start"
+        alignContent="flex-start"
+        alignItems="flex-end"
+        direction="column"
+        wrap="nowrap"
+      >
+        {data?.map((message, index) => {
+          return (
+            (<Grid key={index}>
+              <Grid
+                container
+                justifyContent="flex-start"
+                alignContent="flex-start"
+                alignItems="flex-end"
+                direction="column"
+              >
+                <Grid>
+                  <Chip
+                    style={{height: "auto", paddingTop: "-10px"}}
+                    color="primary"
+                    label={
+                      <Typography align="left">
+                        {message.content}
+                      </Typography>
+                    }/>
+                </Grid>
+                <Grid>
+                  <Typography
+                    variant="caption"
+                    display="block"
 
-const ChatContent = ({ data, handleMessageSend }) => {
-  const classes = useStyles();
-  const [comment, setComment] = React.useState("");
-
-  return (
-    <React.Fragment>
-      <Box paddingLeft={2}>
-        <Grid
-          container
-          justify="flex-start"
-          alignContent="flex-start"
-          alignItems="flex-end"
-          direction="column"
-          className={classes.padding}
-          wrap="nowrap"
-        >
-          {data?.map((message, index) => {
-            return (
-              <Grid item key={index}>
-                <Grid
-                  container
-                  justify="flex-start"
-                  alignContent="flex-start"
-                  alignItems="flex-end"
-                  direction="column"
-                >
-                  <Grid item>
-                    <Chip
-                      style={{ height: "auto", paddingTop: "-10px" }}
-                      color="primary"
-                      variant="default"
-                      label={
-                        <Typography align="left" className={classes.message}>
-                          {message.content}
-                        </Typography>
-                      }
-                    />
-                  </Grid>
-                  <Grid item>
-                    <Typography
-                      variant="caption"
-                      display="block"
-                      className={classes.date}
-                      color="textSecondary"
-                    >
-                      {moment
-                        .utc(message.creation_time)
-                        .local()
-                        .format("ddd Do MMM HH:mm:ss")}{" "}
-                      by {message.by}
-                    </Typography>
-                  </Grid>
+                    color="textSecondary"
+                  >
+                    {moment
+                      .utc(message.creation_time)
+                      .local()
+                      .format("ddd Do MMM HH:mm:ss")}{" "}
+                    by {message.by}
+                  </Typography>
                 </Grid>
               </Grid>
-            );
-          })}
+            </Grid>)
+          );
+        })}
+      </Grid>
+      <Grid
+        container
+        justifyContent="flex-start"
+        alignContent="flex-start"
+        alignItems="center"
+
+      >
+        <Grid size={10}>
+          <TextField
+            id="message"
+            label="Add comment"
+            multiline
+            fullWidth
+            value={comment}
+            onChange={(e) => setComment(e.target.value)}
+          />
         </Grid>
-        <Grid
-          container
-          justify="flex-start"
-          alignContent="flex-start"
-          alignItems="center"
-          className={classes.padding}
-        >
-          <Grid item xs={10}>
-            <TextField
-              id="message"
-              label="Add comment"
-              multiline
-              variant="outlined"
+        <Grid size={2}>
+          <Box paddingLeft={1}>
+            <Button
+              variant="contained"
               fullWidth
-              value={comment}
-              onChange={(e) => setComment(e.target.value)}
-            />
-          </Grid>
-          <Grid item xs={2}>
-            <Box paddingLeft={1}>
-              <Button
-                variant="contained"
-                fullWidth
-                color="secondary"
-                onClick={(e) => {
-                  handleMessageSend(comment);
-                  setComment("");
-                }}
-              >
-                <Send />
-              </Button>
-            </Box>
-          </Grid>
+              color="secondary"
+              onClick={(e) => {
+                handleMessageSend(comment);
+                setComment("");
+              }}
+            >
+              <Send/>
+            </Button>
+          </Box>
         </Grid>
-      </Box>
-    </React.Fragment>
+      </Grid>
+    </Box>
   );
 };
 
-const ChatWidget = ({ data, handleMessageSend }) => {
-  const classes = useStyles();
+const ChatWidget = ({data, handleMessageSend}) => {
   const [open, setOpen] = useState(false);
 
   const handleChange = (reason) => {
@@ -144,11 +106,11 @@ const ChatWidget = ({ data, handleMessageSend }) => {
   };
   // TODO replace with builtin speeddial from MUI
   return (
-    <div className={classes.chatPosition}>
-      <Badge color="secondary" overlap="circle" badgeContent={data?.length}>
-        <Grid container className={classes.shape}>
-          <IconButton onClick={handleChange}>
-            <Comment style={{ color: "white" }} />
+    (<div>
+      <Badge color="secondary" overlap="circular" badgeContent={data?.length}>
+        <Grid container>
+          <IconButton onClick={handleChange} size="large">
+            <Comment style={{color: "white"}}/>
           </IconButton>
         </Grid>
       </Badge>
@@ -156,12 +118,12 @@ const ChatWidget = ({ data, handleMessageSend }) => {
         <ChatContent
           data={data}
           handleMessageSend={handleMessageSend}
-          classes={classes}
+
         />
       </Drawer>
-    </div>
+    </div>)
   );
 };
 
 export default ChatWidget;
-export { ChatWidget, ChatContent };
+export {ChatWidget, ChatContent};
