@@ -1,12 +1,13 @@
 import React from "react";
-import { Grid, Link, Typography } from "@material-ui/core";
+import { Link, Typography } from "@mui/material";
+import Grid from "@mui/material/Grid2";
 import {
   get as apiGet,
   handle_http_errors,
   showResponse,
 } from "../../utils/fetchUtils";
 
-const Footer = ({ classes }) => {
+const Footer = () => {
   const [repoData, setRepoData] = React.useState([]);
   const [apiVersion, setApiVersion] = React.useState("N/A");
 
@@ -28,8 +29,8 @@ const Footer = ({ classes }) => {
   }, []);
 
   return (
-    <Grid container>
-      <Grid item className={classes.paddingTop} xs={12}>
+    (<Grid container>
+      <Grid size={12}>
         <Typography
           color="textSecondary"
           variant="caption"
@@ -48,7 +49,7 @@ const Footer = ({ classes }) => {
         </Typography>
       </Grid>
       {!process.env.REACT_APP_PUBLIC_BUILD ? (
-        <Grid item xs={12}>
+        <Grid size={12}>
           <Typography
             color="textSecondary"
             variant="caption"
@@ -66,7 +67,7 @@ const Footer = ({ classes }) => {
       ) : (
         ""
       )}
-    </Grid>
+    </Grid>)
   );
 };
 
