@@ -1,5 +1,5 @@
-import { TextField } from "@material-ui/core";
-import { Autocomplete } from "@material-ui/lab";
+import { TextField } from "@mui/material";
+import { Autocomplete } from "@mui/material";
 import { MapDetail } from "./map-details";
 
 export function MapAutocomplete({ options, onChange, ...props }) {
@@ -16,9 +16,14 @@ export function MapAutocomplete({ options, onChange, ...props }) {
       renderInput={(params) => (
         <TextField {...params} variant="outlined" label="Select maps" />
       )}
-      renderOption={(option) => (
-        <MapDetail mapLayer={option} />
-      )}
+      renderOption={(props, option) => {
+        const { key, ...optionProps } = props;
+        return (
+          <li key={key} {...optionProps}>
+            <MapDetail mapLayer={option} />
+          </li>
+        );
+      }}
       {...props}
     />
   );
