@@ -1,25 +1,24 @@
-import * as React from "react";
-
-import makeStyles from "@material-ui/core/styles/makeStyles";
-import ListItem from "@material-ui/core/ListItem";
-import ListItemAvatar from "@material-ui/core/ListItemAvatar";
-import ListItemText from "@material-ui/core/ListItemText";
-import { ListItemSecondaryAction, createStyles } from "@material-ui/core";
+import ListItem from "@mui/material/ListItem";
+import ListItemAvatar from "@mui/material/ListItemAvatar";
+import ListItemText from "@mui/material/ListItemText";
+import { ListItemButton } from "@mui/material";
 import { MapAvatar, MapDescription } from "./map-details";
 
-const useStyles = makeStyles((theme) =>
-  createStyles({
-    root: {
-      borderBottom: `1px solid ${theme.palette.divider}`,
-    },
-  })
+const ListItemComponent = ({ children, secondaryAction, asButton, ...props }) => asButton ? (
+  <ListItem secondaryAction={secondaryAction} disablePadding>
+    <ListItemButton role={undefined} disableRipple dense {...props}>
+      {children}
+    </ListItemButton>
+  </ListItem>
+) : (
+  <ListItem secondaryAction={secondaryAction} {...props}>
+    {children}
+  </ListItem>
 );
 
-export function MapListItem({ mapLayer, primary, secondary, renderAction, ...props }) {
-  const classes = useStyles();
-
+export function MapListItem({ mapLayer, primary, secondary, secondaryAction, asButton, ...props }) {
   return (
-    <ListItem className={classes.root} {...props}>
+    <ListItemComponent secondaryAction={secondaryAction} asButton={asButton} {...props}>
       <ListItemAvatar>
         <MapAvatar mapLayer={mapLayer} />
       </ListItemAvatar>
@@ -27,11 +26,6 @@ export function MapListItem({ mapLayer, primary, secondary, renderAction, ...pro
         primary={primary ?? mapLayer.map.pretty_name}
         secondary={secondary ?? <MapDescription mapLayer={mapLayer} />}
       />
-      {renderAction && (
-        <ListItemSecondaryAction>
-          {renderAction(mapLayer)}
-        </ListItemSecondaryAction>
-      )}
-    </ListItem>
+    </ListItemComponent>
   );
 }
