@@ -1,31 +1,30 @@
-import React from "react";
-import Button from "@material-ui/core/Button";
-import Chip from "@material-ui/core/Chip";
-import ButtonGroup from "@material-ui/core/ButtonGroup";
-import TextField from "@material-ui/core/TextField";
-import _ from "lodash";
-import Badge from "@material-ui/core/Badge";
-import Menu from "@material-ui/core/Menu";
-import MenuItem from "@material-ui/core/MenuItem";
-import ArrowDropDownIcon from "@material-ui/icons/ArrowDropDown";
+import Button from "@mui/material/Button";
+import Chip from "@mui/material/Chip";
+import ButtonGroup from "@mui/material/ButtonGroup";
+import TextField from "@mui/material/TextField";
+import Badge from "@mui/material/Badge";
+import Menu from "@mui/material/Menu";
+import MenuItem from "@mui/material/MenuItem";
+import ArrowDropDownIcon from "@mui/icons-material/ArrowDropDown";
 import "react-toastify/dist/ReactToastify.css";
-import Dialog from "@material-ui/core/Dialog";
-import DialogActions from "@material-ui/core/DialogActions";
-import DialogContent from "@material-ui/core/DialogContent";
-import DialogTitle from "@material-ui/core/DialogTitle";
+import Dialog from "@mui/material/Dialog";
+import DialogActions from "@mui/material/DialogActions";
+import DialogContent from "@mui/material/DialogContent";
+import DialogTitle from "@mui/material/DialogTitle";
 import { Map } from "immutable";
-import FlagOutlinedIcon from "@material-ui/icons/FlagOutlined";
+import FlagOutlinedIcon from "@mui/icons-material/FlagOutlined";
 import TextHistory from "../textHistory";
-import Autocomplete from "@material-ui/lab/Autocomplete";
-import FormControlLabel from "@material-ui/core/FormControlLabel";
-import Checkbox from "@material-ui/core/Checkbox";
-import { getSharedMessages } from "../../utils/fetchUtils";
-import { Grid } from "@material-ui/core";
-import Tooltip from "@material-ui/core/Tooltip";
-import MessageIcon from "@material-ui/icons/Message";
-import VisibilityIcon from "@material-ui/icons/Visibility";
-import BlockIcon from "@material-ui/icons/Block";
-import StarIcon from "@material-ui/icons/Star";
+import Autocomplete from '@mui/material/Autocomplete';
+import FormControlLabel from "@mui/material/FormControlLabel";
+import Checkbox from "@mui/material/Checkbox";
+import { getSharedMessages } from "@/utils/fetchUtils";
+import Grid from "@mui/material/Grid2";
+import Tooltip from "@mui/material/Tooltip";
+import MessageIcon from "@mui/icons-material/Message";
+import VisibilityIcon from "@mui/icons-material/Visibility";
+import BlockIcon from "@mui/icons-material/Block";
+import StarIcon from "@mui/icons-material/Star";
+import {Component, Fragment, useState} from "react";
 
 const Duration = ({
   durationNumber,
@@ -34,7 +33,7 @@ const Duration = ({
   onMultiplierChange,
 }) => (
   <Grid container spacing={1}>
-    <Grid item>
+    <Grid>
       <TextField
         style={{ minWidth: "200px" }}
         label="TempBan Duration number"
@@ -45,7 +44,7 @@ const Duration = ({
         onChange={(event) => onNumberChange(event.target.value)}
       />
     </Grid>
-    <Grid item>
+    <Grid>
       <TextField
         style={{ minWidth: "200px" }}
         select
@@ -71,7 +70,7 @@ const Duration = ({
   </Grid>
 );
 
-class ReasonDialog extends React.Component {
+class ReasonDialog extends Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -135,7 +134,7 @@ class ReasonDialog extends React.Component {
     const actionType = open.actionType;
     const playerName = open.player;
     return (
-      <Dialog open={open} aria-labelledby="form-dialog-title">
+      (<Dialog open={open} aria-labelledby="form-dialog-title">
         <DialogTitle id="form-dialog-title">
           {this.mapActionToText(actionType, playerName)}
         </DialogTitle>
@@ -150,10 +149,9 @@ class ReasonDialog extends React.Component {
               <TextField
                 {...params}
                 multiline
-                rows={4}
-                rowsMax={10}
+                minRows={4}
+                maxRows={10}
                 label={actionType === "message_player" ? "Message" : "Reason"}
-                variant="outlined"
                 margin="dense"
                 helperText="The message that will be displayed to the player. A message is mandatory"
               />
@@ -163,13 +161,12 @@ class ReasonDialog extends React.Component {
           {open.actionType !== "message_player" ? (
             <TextField
               multiline
-              rows={4}
-              rowsMax={10}
+              minRows={4}
+              maxRows={10}
               fullWidth
               value={comment}
               onChange={(e) => this.setState({ comment: e.target.value })}
               label="Comment"
-              variant="outlined"
               margin="dense"
               helperText="A comment that will NOT be displayed to the player"
             />
@@ -231,7 +228,7 @@ class ReasonDialog extends React.Component {
             Confirm
           </Button>
         </DialogActions>
-      </Dialog>
+      </Dialog>)
     );
   }
 }
@@ -248,8 +245,8 @@ const PlayerActions = ({
   onBlacklist,
   onVipDialogOpen,
 }) => {
-  const [anchorEl, setAnchorEl] = React.useState(null);
-  const [isOpen, setOpen] = React.useState(false);
+  const [anchorEl, setAnchorEl] = useState(null);
+  const [isOpen, setOpen] = useState(false);
   const handleClick = (event) => {
     setAnchorEl(event.currentTarget);
     setOpen(true);
@@ -275,7 +272,7 @@ const PlayerActions = ({
   const show = Math.min(displayCount, actions.length);
 
   return (
-    <React.Fragment>
+    <Fragment>
       <ButtonGroup
         size={size}
         disabled={disableAll}
@@ -289,8 +286,6 @@ const PlayerActions = ({
         {show > 1 ? (
           <Tooltip title="Watch Player">
             <Button
-              color={isWatched ? "primary" : "default"}
-              variant={isWatched ? "contained" : "outlined"}
               size="small"
               onClick={() =>
                 handleAction(isWatched ? "unwatch_player" : "watch_player")
@@ -310,7 +305,6 @@ const PlayerActions = ({
           >
             <Badge
               size="small"
-              color="primary"
               max={9}
               badgeContent={penaltyCount.get(
                 remap_penalties[actions[idx][0]],
@@ -370,7 +364,6 @@ const PlayerActions = ({
               }
             >
               <VisibilityIcon
-                color={isWatched ? "primary" : "default"}
                 fontSize="small"
               />
             </MenuItem>
@@ -400,7 +393,7 @@ const PlayerActions = ({
       ) : (
         ""
       )}
-    </React.Fragment>
+    </Fragment>
   );
 };
 
