@@ -6,19 +6,18 @@ import {
   ListItemAvatar,
   ListItemText,
   Typography,
-} from "@material-ui/core";
+} from "@mui/material";
 import { Link as RouterLink } from "react-router-dom";
-import React from "react";
 import { List, Map } from "immutable";
-import Tooltip from "@material-ui/core/Tooltip";
-import FileCopyIcon from "@material-ui/icons/FileCopy";
-import AnnouncementIcon from "@material-ui/icons/Announcement";
-import ListItemSecondaryAction from "@material-ui/core/ListItemSecondaryAction";
-import KeyboardArrowUpIcon from "@material-ui/icons/KeyboardArrowUp";
-import KeyboardArrowDownIcon from "@material-ui/icons/KeyboardArrowDown";
-import { pure } from "recompose";
+import Tooltip from "@mui/material/Tooltip";
+import FileCopyIcon from "@mui/icons-material/FileCopy";
+import AnnouncementIcon from "@mui/icons-material/Announcement";
+import ListItemSecondaryAction from "@mui/material/ListItemSecondaryAction";
+import KeyboardArrowUpIcon from "@mui/icons-material/KeyboardArrowUp";
+import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
 import { getName } from "country-list";
 import makePlayerProfileUrl from "../../../utils/makePlayerProfileUrl";
+import {Fragment, useState} from "react";
 
 export const getCountry = (country) => {
   if (country === "" || country === null) {
@@ -34,8 +33,8 @@ export const getCountry = (country) => {
   );
 };
 
-export const PlayerHeader = pure(({ classes, player }) => {
-  const [showAll, setShowAll] = React.useState(false);
+export const PlayerHeader = ({ player }) => {
+  const [showAll, setShowAll] = useState(false);
   const hasMultipleName = player.get("names") && player.get("names").size > 1;
 
   const playerNames = player.get("names", null)
@@ -70,7 +69,7 @@ export const PlayerHeader = pure(({ classes, player }) => {
       </ListItemAvatar>
       <ListItemText
         primary={
-          <React.Fragment>
+          <Fragment>
             {showAll ? (
               <Typography variant="body1">
                 {hasMultipleName ? (
@@ -102,13 +101,13 @@ export const PlayerHeader = pure(({ classes, player }) => {
                 {namesByMatch.get(0, firstName?.get("name"))} {getCountry(country)}
               </Typography>
             )}
-          </React.Fragment>
+          </Fragment>
         }
         secondary={
           <Link
             color="inherit"
             component={RouterLink}
-            to={`/player/${player.get("player_id")}`}
+            to={`/records/players/${player.get("player_id")}`}
           >
             {player.get("player_id")}
           </Link>
@@ -182,4 +181,4 @@ export const PlayerHeader = pure(({ classes, player }) => {
       </ListItemSecondaryAction>
     </ListItem>
   );
-});
+};
