@@ -1,9 +1,8 @@
-import { Grid } from "@material-ui/core";
-import React from "react";
-import Tooltip from "@material-ui/core/Tooltip";
+import Grid from "@mui/material/Grid2";
+import Tooltip from "@mui/material/Tooltip";
 import moment from "moment";
 
-export const PlayerSighthings = ({ classes, player }) => {
+export const PlayerSighthings = ({ player }) => {
   const first_seen = moment(player.get("first_seen_timestamp_ms"));
   const last_seen = moment(player.get("last_seen_timestamp_ms"));
 
@@ -31,14 +30,13 @@ export const PlayerSighthings = ({ classes, player }) => {
   }
 
   return (
-    <Grid
-      container
-      justify="space-between"
-      spacing={0}
-      className={classes.noPaddingMargin}
-    >
+    (<Grid
+        container
+        justifyContent="space-between"
+        spacing={0}
+      >
       {vip_expiration ? (
-        <Grid item xs={12}>
+        <Grid size={12}>
           <Tooltip title={vip_expiration.format("LLLL")} arrow>
             {vipDisplay}
           </Tooltip>
@@ -46,21 +44,20 @@ export const PlayerSighthings = ({ classes, player }) => {
       ) : (
         ""
       )}
-
-      <Grid item xs={6}>
+      <Grid size={6}>
         <Tooltip title={first_seen.format("LLLL")} arrow>
           <small>
             First seen {moment.duration(now.diff(first_seen)).humanize()} ago
           </small>
         </Tooltip>
       </Grid>
-      <Grid item xs={6}>
+      <Grid size={6}>
         <Tooltip title={last_seen.format("LLLL")} arrow>
           <small>
             Last seen {moment.duration(now.diff(last_seen)).humanize()} ago
           </small>
         </Tooltip>
       </Grid>
-    </Grid>
+    </Grid>)
   );
 };
