@@ -1,10 +1,17 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import "./index.css";
 import App from "./App";
+import "./index.css";
 import * as serviceWorker from "./serviceWorker";
+import { createRoot } from "react-dom/client";
+import '@fontsource/roboto/300.css';
+import '@fontsource/roboto/400.css';
+import '@fontsource/roboto/500.css';
+import '@fontsource/roboto/700.css';
 
-ReactDOM.render(<App />, document.getElementById("root"));
+const app = document.getElementById('root');
+
+const root = createRoot(app);
+
+root.render(<App />);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
