@@ -1,5 +1,3 @@
-import faker from "faker";
-
 // a little function to help us with reordering the result
 export const reorder = (list, startIndex, endIndex) => {
   const result = Array.from(list);
@@ -8,16 +6,6 @@ export const reorder = (list, startIndex, endIndex) => {
 
   return result;
 };
-
-export const getItems = (count) =>
-  Array.from({ length: count }, (v, k) => k).map(
-    (k) => faker.commerce.productName()
-    //{
-    //id: `Item ${k + 1}`,
-    // primary: faker.commerce.productName(),
-    //secondary: faker.company.catchPhrase()
-    //}
-  );
 
 export const unifiedGamemodeName = (modeName) => {
   switch (modeName) {
@@ -58,5 +46,5 @@ export const generateObjectivesGrid = (orientation) => {
     });
 };
 
-export const getMapLayerImageSrc = (mapLayer) => `maps/icons/${mapLayer.image_name}`;
-export const getTacMapImageSrc = (mapLayer) => `tac-maps/${mapLayer.map.id}.webp`;
+export const getMapLayerImageSrc = (mapLayer) => `/maps/icons/${mapLayer.image_name}`;
+export const getTacMapImageSrc = (mapLayer) => `/tac-maps/${mapLayer.map.id}.webp`;
