@@ -1,3 +1,3 @@
-import PlayerInfo from "./PlayerInfo";
+import PlayerInfo from "../../pages/records/players/[playerId]";
 
 export default PlayerInfo;
