@@ -1,22 +1,4 @@
-import _ from "lodash";
-
 const PREFIX = "autocomplete_";
-
-const getAllNamespaces = () => {
-  let namespaces = ["punishments", "welcome", "broadcast", "watchlist"];
-
-  return namespaces.filter((v) => {
-    if (!v || v.includes("undefined")) {
-      return false;
-    }
-    const texts = new TextHistory(v).getTexts();
-
-    if (texts.length === 0) {
-      return false;
-    }
-    return true;
-  });
-};
 
 class TextHistory {
   constructor(namespace) {
@@ -38,12 +20,6 @@ class TextHistory {
     return texts;
   }
 
-  deleteTextByIdx(index) {
-    const texts = this.getTexts();
-    texts.splice(index, 1);
-    localStorage.setItem(this.namespace, JSON.stringify(texts));
-  }
-
   saveText(text, sharedMessages = []) {
     if (!text || sharedMessages.includes(text)) {
       return;
@@ -59,4 +35,4 @@ class TextHistory {
 }
 
 export default TextHistory;
-export { TextHistory, getAllNamespaces };
+export { TextHistory };
