@@ -1,11 +1,4 @@
-import React from "react";
-import { Emoji, getEmojiDataFromNative } from "emoji-mart";
-import data from "emoji-mart/data/all.json";
 
-export function getEmojiFlag(flag, size = 22) {
-  const emo = getEmojiDataFromNative(flag, "apple", data);
-  if (emo) {
-    return <Emoji emoji={emo} set="apple" size={size} />;
-  }
-  return flag;
+export function getEmojiFlag(flag) {
+  return flag
 }
